w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/shadows/2d.shadow.pattern.alpha-expected.txt

2d.shadow.pattern.alpha
Shadows are drawn correctly for partially-transparent fill patterns
Actual output:
Expected output:


FAIL Shadows are drawn correctly for partially-transparent fill patterns assert_approx_equals: Red channel of the pixel at (50, 25) expected 127 +/- 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:47:25
http://localhost:8800/2dcontext/shadows/2d.shadow.pattern.alpha.html:30:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
